renamed src to forager
